feat: refactor audio search modal
